Conversion work to use a workspace with multiple libraries.
Some work done to standardise code as part of the migration.

More work to be done to DRY up some code and create common methods etc.

Code builds with some warnings due to ongoing DRYing up.

As yet untested.
