+++ Cargo.toml
@@ -1,31 +1,8 @@
-[package]
-name = "acars_router"
-version = "1.0.4"
-edition = "2021"
-authors = ["Fred Clausen", "Mike Nye"]
-description = "ACARS Router: A Utility to ingest ACARS/VDLM2 from many sources, process, and feed out to many consumers."
-documentation = "https://github.com/sdr-enthusiasts/acars_router"
-homepage = "https://github.com/sdr-enthusiasts/acars_router"
-repository = "https://github.com/sdr-enthusiasts/acars_router"
-readme = "README.md"
-license = "MIT"
-
-[dependencies]
-clap = { version = "3.2.15", features = ["derive", "env"] }
-log = "0.4.17"
-env_logger = "0.9.0"
-chrono = "0.4.19"
-tokio = { version = "1.20.1", features = ["full", "tracing"] }
-tokio-util = { version = "0.7.3", features = ["full"] }
-tokio-stream = "0.1.9"
-serde = { version = "1.0.140", features = ["derive"] }
-serde_json = "1.0.82"
-derive-getters = "0.2.0"
-zmq = "0.9.2"
-tmq = "0.3.2"
-futures = "0.3.21"
-failure = "0.1.8"
-stubborn-io = "0.3.1"
-acars_vdlm2_parser = { git = "https://github.com/jcdeimos/acars_vdlm2_parser", version = "0.1.2" }
-async-trait = "0.1.56"
-nom = "7.1.1"
+[workspace]
+members = [
+    "rust/bin/acars_router",
+    "rust/libraries/acars_config",
+    "rust/libraries/acars_data_processing",
+    "rust/libraries/acars_logging",
+    "rust/libraries/acars_connection_manager"
+]
