+++ rust/bin/acars_router/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "acars_router"
+version = "1.0.4"
+edition = "2021"
+authors = ["Fred Clausen", "Mike Nye"]
+description = "ACARS Router: A Utility to ingest ACARS/VDLM2 from many sources, process, and feed out to many consumers."
+documentation = "https://github.com/sdr-enthusiasts/acars_router"
+homepage = "https://github.com/sdr-enthusiasts/acars_router"
+repository = "https://github.com/sdr-enthusiasts/acars_router"
+readme = "README.md"
+license = "MIT"
+
+[dependencies]
+log = "0.4.17"
+tokio = { version = "1.20.1", features = ["full", "tracing"] }
+#tokio-util = { version = "0.7.3", features = ["full"] }
+#tokio-stream = "0.1.9"
+#futures = "0.3.21"
+chrono = "0.4.19"
+serde = { version = "1.0.140", features = ["derive"] }
+serde_json = "1.0.82"
+#derive-getters = "0.2.0"
+#zmq = "0.9.2"
+#tmq = "0.3.2"
+failure = "0.1.8"
+#stubborn-io = "0.3.1"
+#async-trait = "0.1.56"
+#acars_vdlm2_parser = { git = "https://github.com/jcdeimos/acars_vdlm2_parser", version = "0.1.2" }
+acars_config = { path = "../../libraries/acars_config" }
+acars_logging = { path = "../../libraries/acars_logging" }
+acars_connection_manager = { path = "../../libraries/acars_connection_manager" }
