+++ rust/bin/acars_router/src/main.rs
@@ -1,0 +1,40 @@
+// Copyright (c) Mike Nye, Fred Clausen
+//
+// Licensed under the MIT license: https://opensource.org/licenses/MIT
+// Permission is granted to use, copy, modify, and redistribute the work.
+// Full license information available in the project LICENSE file.
+//
+
+#[macro_use] extern crate log;
+extern crate failure;
+extern crate serde_json;
+extern crate serde;
+extern crate chrono;
+extern crate acars_config;
+extern crate acars_logging;
+extern crate acars_connection_manager;
+
+use std::error::Error;
+use std::process;
+use acars_config::clap::Parser;
+use acars_config::Input;
+use acars_config::sanity_checker::check_config_option_sanity;
+use acars_connection_manager::service_init::start_processes;
+use acars_logging::SetupLogging;
+
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn Error>> {
+    let args: Input = Input::parse();
+    args.verbose.enable_logging();
+    match check_config_option_sanity(&args) {
+        Ok(_) => {
+            trace!("Config options are sane");
+        }
+        Err(e) => {
+            error!("{}", e);
+            process::exit(1);
+        }
+    }
+    start_processes(args).await;
+    Ok(())
+}
