+++ rust/libraries/acars_config/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "acars_config"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+clap = { version = "3.2.15", features = ["derive", "env"] }
+log = "0.4.17"
+acars_logging = { path = "../acars_logging" }
