+++ rust/libraries/acars_config/src/sanity_checker.rs
@@ -7,131 +7,131 @@
 
 // File to verify the sanity of config options
 
-use crate::config_options::Input;
+use crate::Input;
 use std::net::SocketAddr;
 use std::str::FromStr;
 
 pub fn check_config_option_sanity(config_options: &Input) -> Result<(), String> {
     let mut is_input_sane = true;
-
+    
     // We don't have to verify the log level because the config_options module
     // Sanitizes that to a sane value
-
+    
     // Also, any input option set to be *only* numeric/u64 (ie, skew_window)
     // Will always be a valid number because the Clap parser will die if the input is bad
-
+    
     if !check_ports_are_valid(
         &config_options.listen_udp_acars,
         "AR_LISTEN_UDP_ACARS/--listen-udp-acars",
     ) {
         is_input_sane = false;
     }
-
+    
     if !check_ports_are_valid(
         &config_options.listen_tcp_acars,
         "AR_LISTEN_TCP_ACARS/--listen-tcp-acars",
     ) {
         is_input_sane = false;
     }
-
+    
     if !check_ports_are_valid_with_host(
         &config_options.receive_tcp_acars,
         "AR_RECEIVE_TCP_ACARS/--receive-tcp-acars",
     ) {
         is_input_sane = false;
     }
-
+    
     if !check_ports_are_valid(
         &config_options.listen_udp_vdlm2,
         "AR_LISTEN_UDP_VDLM2/--listen-udp-vdlm2",
     ) {
         is_input_sane = false;
     }
-
+    
     if !check_ports_are_valid(
         &config_options.listen_tcp_vdlm2,
         "AR_LISTEN_TCP_VDLM2/--listen-tcp-vdlm2",
     ) {
         is_input_sane = false;
     }
-
+    
     if !check_ports_are_valid_with_host(
         &config_options.receive_tcp_vdlm2,
         "AR_RECEIVE_TCP_VDLM2/--receive-tcp-vdlm2",
     ) {
         is_input_sane = false;
     }
-
+    
     if !check_ports_are_valid_with_host(
         &config_options.send_udp_acars,
         "AR_SEND_UDP_ACARS/--send-udp-acars",
     ) {
         is_input_sane = false;
     }
-
+    
     if !check_ports_are_valid_with_host(
         &config_options.send_udp_vdlm2,
         "AR_SEND_UDP_VDLM2/--send-udp-vdlm2",
     ) {
         is_input_sane = false;
     }
-
+    
     if !check_ports_are_valid_with_host(
         &config_options.send_tcp_acars,
         "AR_SEND_TCP_ACARS/--send-tcp-acars",
     ) {
         is_input_sane = false;
     }
-
+    
     if !check_ports_are_valid_with_host(
         &config_options.send_tcp_vdlm2,
         "AR_SEND_TCP_VDLM2/--send-tcp-vdlm2",
     ) {
         is_input_sane = false;
     }
-
+    
     if !check_ports_are_valid_with_host(
         &config_options.receive_zmq_vdlm2,
         "AR_RECEIVE_ZMQ_VDLM2/--receive-zmq-vdlm2",
     ) {
         is_input_sane = false;
     }
-
+    
     if !check_ports_are_valid_with_host(
         &config_options.receive_zmq_acars,
         "AR_RECEIVE_ZMQ_ACARS/--receive-zmq-acars",
     ) {
         is_input_sane = false;
     }
-
+    
     if !check_ports_are_valid(
         &config_options.serve_tcp_acars,
         "AR_SERVE_TCP_ACARS/--serve-tcp-acars",
     ) {
         is_input_sane = false;
     }
-
+    
     if !check_ports_are_valid(
         &config_options.serve_tcp_vdlm2,
         "AR_SERVE_TCP_VDLM2/--serve-tcp-vdlm2",
     ) {
         is_input_sane = false;
     }
-
+    
     if !check_ports_are_valid(
         &config_options.serve_zmq_acars,
         "AR_SERVE_ZMQ_ACARS/--serve-zmq-acars",
     ) {
         is_input_sane = false;
     }
-
+    
     if !check_ports_are_valid(
         &config_options.serve_zmq_vdlm2,
         "AR_SERVE_ZMQ_VDLM2/--serve-zmq-vdlm2",
     ) {
         is_input_sane = false;
     }
-
+    
     match is_input_sane {
         true => Ok(()),
         false => Err("Config option sanity check failed".to_string()),
@@ -158,7 +158,7 @@ fn check_ports_are_valid(option_ports: &Option<Vec<u16>>, name: &str) -> bool {
 
 fn check_ports_are_valid_with_host(socket_addresses: &Option<Vec<String>>, name: &str) -> bool {
     let mut is_input_sane = true;
-
+    
     if let Some(sockets) = socket_addresses {
         if sockets.is_empty() {
             error!("{} has been provided, but there are no socket addresses",name);
@@ -227,7 +227,7 @@ fn check_ports_are_valid_with_host(socket_addresses: &Option<Vec<String>>, name:
 #[cfg(test)]
 mod test {
     use super::*;
-
+    
     #[test]
     fn test_check_ports_are_valid_with_host() {
         let valid_hosts: Option<Vec<String>> = Some(vec![
@@ -255,7 +255,7 @@ mod test {
         assert_eq!(invalid_hosts_tests, false);
         assert_eq!(empty_host_vec_test, false);
     }
-
+    
     #[test]
     fn test_check_ports_are_valid() {
         let valid_ports: Option<Vec<u16>> = Some(vec![1, 8008, 65535]);
