+++ rust/libraries/acars_connection_manager/src/packet_handler.rs
@@ -56,27 +56,27 @@ impl PacketHandler {
             reassembly_window,
         }
     }
-
+    
     pub async fn attempt_message_reassembly(&self, new_message_string: String, peer: SocketAddr) -> Option<AcarsVdlm2Message> {
         // FIXME: This is a hack to get this concept working. Ideally we aren't cleaning the queue while
         // Processing a message
         self.clean_queue().await;
         // FIXME: Ideally this entire function should not lock the mutex all the time
-
+        
         if let Ok(msg) = new_message_string.decode_message() {
             self.queue.lock().await.remove(&peer);
             return Some(msg);
         }
-
+        
         let mut output_message: Option<AcarsVdlm2Message> = None;
         let mut message_for_peer = String::new();
         let mut old_time: Option<f64> = None; // Save the time of the first message for this peer
-
+        
         // TODO: Handle message reassembly for out of sequence messages
         // TODO: Handle message reassembly for a peer where the peer is sending multiple fragmented messages
         // Maybe on those two? This could get really tricky to know if the message we've reassembled is all the same message
         // Because we could end up in a position where the packet splits in the same spot and things look right but the packets belong to different messages
-
+        
         if self.queue.lock().await.contains_key(&peer) {
             info!("[UDP SERVER: {}] Message received from {} is being reassembled", self.name, peer);
             let (time, message_to_test) = self.queue.lock().await.get(&peer).unwrap().clone();
@@ -94,7 +94,7 @@ impl PacketHandler {
                 }
             };
         }
-
+        
         match output_message {
             Some(_) => {
                 self.queue.lock().await.remove(&peer);
@@ -105,11 +105,11 @@ impl PacketHandler {
                 if message_for_peer.is_empty() {
                     message_for_peer = new_message_string;
                 }
-
+                
                 // We want the peer's message queue to expire once the FIRST message received from the peer is older
                 // than the reassembly window. Therefore we use the old_time we grabbed from the queue above, or if it's the first
                 // message we get the current time.
-
+                
                 let message_queue_time = match old_time {
                     Some(t) => t,
                     None => match SystemTime::now().duration_since(UNIX_EPOCH) {
@@ -117,35 +117,35 @@ impl PacketHandler {
                         Err(_) => 0.0,
                     },
                 };
-
+                
                 self.queue.lock().await.insert(peer, (message_queue_time, message_for_peer));
             }
         }
-
+        
         output_message
     }
-
+    
     pub async fn clean_queue(&self) {
         let current_time = match SystemTime::now().duration_since(UNIX_EPOCH) {
             Ok(n) => n.as_secs_f64(),
             Err(_) => 0.0,
         };
-
+        
         if current_time == 0.0 {
             error!("[UDP SERVER: {}] Error getting current time", self.name);
             return;
         }
-
+        
         self.queue.lock().await.retain(|peer, old_messages| {
-                let (time, _) = old_messages;
-                let time_diff = current_time - *time;
-                if time_diff > self.reassembly_window {
-                    debug!("[UDP SERVER {}] Peer {peer} has been idle for {time_diff} seconds, removing from queue", self.name);
-                    false
-                } else {
-                    debug!("[UDP SERVER {}] Peer {peer} has been idle for {time_diff} seconds, keeping in queue", self.name);
-                    true
-                }
-            });
+            let (time, _) = old_messages;
+            let time_diff = current_time - *time;
+            if time_diff > self.reassembly_window {
+                debug!("[UDP SERVER {}] Peer {peer} has been idle for {time_diff} seconds, removing from queue", self.name);
+                false
+            } else {
+                debug!("[UDP SERVER {}] Peer {peer} has been idle for {time_diff} seconds, keeping in queue", self.name);
+                true
+            }
+        });
     }
 }
