+++ rust/libraries/acars_logging/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "acars_logging"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+log = "0.4.17"
+env_logger = "0.9.0"
+chrono = "0.4.19"
